chore: upgrade workspace dependencies
