+++ .github/workflows/cli-release.yml
@@ -14,7 +14,7 @@ on:
         required: false
         type: string
   release:
-    types: [released]
+    types: [ released ]
 
 concurrency:
   group: difyctl-release
@@ -114,7 +114,7 @@ jobs:
         run: node scripts/release-naming.mjs github-env >> "$GITHUB_ENV"
 
       - name: Setup Bun
-        uses: oven-sh/setup-bun@4bc047ad259df6fc24a6c9b0f9a0cb08cf17fbe5 # v2.0.2
+        uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2.2.0
         with:
           bun-version-file: cli/.bun-version
