+++ .github/workflows/main-ci.yml
@@ -2,10 +2,10 @@ name: Main CI Pipeline
 
 on:
   pull_request:
-    branches: ["main"]
+    branches: [ "main" ]
   merge_group:
-    branches: ["main"]
-    types: [checks_requested]
+    branches: [ "main" ]
+    types: [ checks_requested ]
 
 permissions:
   actions: write
@@ -15,7 +15,8 @@ permissions:
   statuses: write
 
 concurrency:
-  group: main-ci-${{ github.event.pull_request.number || github.event.merge_group.head_sha || github.run_id }}
+  group: main-ci-${{ github.event.pull_request.number ||
+    github.event.merge_group.head_sha || github.run_id }}
   cancel-in-progress: true
 
 jobs:
@@ -48,7 +49,7 @@ jobs:
       sandbox-runtime-changed: ${{ steps.changes.outputs.sandbox-runtime }}
     steps:
       - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-      - uses: dorny/paths-filter@7b450fff21473bca461d4b92ce414b9d0420d706 # v4.0.2
+      - uses: dorny/paths-filter@ceb8a2b8f2d89434be7ff52d3de7ec3738c5cc9d # v4.0.3
         id: changes
         with:
           filters: |
@@ -154,7 +155,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.api-changed == 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.api-changed == 'true'
     uses: ./.github/workflows/api-tests.yml
     secrets: inherit
 
@@ -163,7 +165,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.api-changed != 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.api-changed != 'true'
     runs-on: depot-ubuntu-24.04
     steps:
       - name: Report skipped API tests
@@ -214,7 +217,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.cli-changed == 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.cli-changed == 'true'
     uses: ./.github/workflows/cli-tests.yml
     secrets: inherit
 
@@ -223,7 +227,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.cli-changed != 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.cli-changed != 'true'
     runs-on: depot-ubuntu-24.04
     steps:
       - name: Report skipped CLI tests
@@ -274,7 +279,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.web-changed == 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.web-changed == 'true'
     uses: ./.github/workflows/web-tests.yml
     secrets: inherit
 
@@ -283,7 +289,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.web-changed != 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.web-changed != 'true'
     runs-on: depot-ubuntu-24.04
     steps:
       - name: Report skipped web tests
@@ -334,7 +341,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.e2e-changed == 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.e2e-changed == 'true'
     uses: ./.github/workflows/web-e2e.yml
     with:
       run-external-runtime: false
@@ -345,7 +353,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.e2e-changed != 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.e2e-changed != 'true'
     runs-on: depot-ubuntu-24.04
     steps:
       - name: Report skipped web full-stack e2e
@@ -397,14 +406,16 @@ jobs:
     if: needs.pre_job.outputs.should_skip != 'true'
     uses: ./.github/workflows/style.yml
     with:
-      base-rev: ${{ github.event.pull_request.base.sha || github.event.merge_group.base_sha }}
+      base-rev: ${{ github.event.pull_request.base.sha ||
+        github.event.merge_group.base_sha }}
 
   vdb-tests-run:
     name: Run VDB Tests
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.vdb-changed == 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.vdb-changed == 'true'
     uses: ./.github/workflows/vdb-tests.yml
 
   vdb-tests-skip:
@@ -412,7 +423,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.vdb-changed != 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.vdb-changed != 'true'
     runs-on: depot-ubuntu-24.04
     steps:
       - name: Report skipped VDB tests
@@ -463,7 +475,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.migration-changed == 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.migration-changed == 'true'
     uses: ./.github/workflows/db-migration-test.yml
 
   db-migration-test-skip:
@@ -471,7 +484,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.migration-changed != 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.migration-changed != 'true'
     runs-on: depot-ubuntu-24.04
     steps:
       - name: Report skipped DB migration tests
@@ -522,7 +536,8 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.sandbox-runtime-changed == 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.sandbox-runtime-changed == 'true'
     uses: ./.github/workflows/sandbox-runtime-tests.yml
     secrets: inherit
 
@@ -531,11 +546,13 @@ jobs:
     needs:
       - pre_job
       - check-changes
-    if: needs.pre_job.outputs.should_skip != 'true' && needs.check-changes.outputs.sandbox-runtime-changed != 'true'
+    if: needs.pre_job.outputs.should_skip != 'true' &&
+      needs.check-changes.outputs.sandbox-runtime-changed != 'true'
     runs-on: depot-ubuntu-24.04
     steps:
       - name: Report skipped sandbox runtime tests
-        run: echo "No sandbox-runtime-related changes detected; skipping sandbox runtime tests."
+        run: echo "No sandbox-runtime-related changes detected; skipping sandbox runtime
+          tests."
 
   sandbox-runtime-tests:
     name: Sandbox Runtime Tests
