+++ package.json
@@ -6,6 +6,9 @@
     "check": "vp check && pnpm lint:eslint",
     "check:fix": "pnpm lint:eslint:fix && vp check --fix",
     "dev": "concurrently -k -n vinext,proxy \"vp run dify-web#dev:vinext\" \"vp run dify-web#dev:proxy\"",
+    "knip": "knip --workspace web",
+    "knip:production": "knip --workspace web --production --include files",
+    "knip:production-unused-check": "node ./web/scripts/check-production-unused-after-knip-fix.mjs",
     "prepare": "vp config",
     "lint:oxlint": "vp lint",
     "lint:oxlint:fix": "vp lint --fix",
@@ -45,6 +48,7 @@
     "eslint-plugin-toml": "catalog:",
     "eslint-plugin-unicorn": "catalog:",
     "eslint-plugin-yml": "catalog:",
+    "knip": "catalog:",
     "typescript": "catalog:",
     "vite": "catalog:",
     "vite-plus": "catalog:"
@@ -59,5 +63,5 @@
   "engines": {
     "node": "^22.22.1"
   },
-  "packageManager": "pnpm@11.17.0"
+  "packageManager": "pnpm@11.20.0"
 }
