+++ pnpm-workspace.yaml
@@ -39,32 +39,32 @@ overrides:
   postcss-selector-parser@>=6.0.0 <6.1.3: 6.1.4
   postcss-selector-parser@>=7.0.0 <7.1.3: 7.1.4
   postcss@<8.5.10: ^8.5.10
-  rollup@>=4.0.0 <4.59.0: 4.62.3
+  rollup@>=4.0.0 <4.59.0: 4.62.4
   safer-buffer: npm:@nolyfill/safer-buffer@^1.0.44
   side-channel: npm:@nolyfill/side-channel@^1.0.44
   solid-js: 1.9.14
   string-width: ~8.2.2
   tar@<=7.5.15: ^7.5.16
-  vite: npm:@voidzero-dev/vite-plus-core@0.2.6
-  ws@>=8.0.0 <8.20.1: ^8.21.1
+  vite: npm:@voidzero-dev/vite-plus-core@0.2.8
+  ws@>=8.0.0 <8.20.1: ^8.21.2
   yaml@>=2.0.0 <2.8.3: 2.9.0
   yauzl@<3.2.1: 3.2.1
 catalog:
-  '@amplitude/analytics-browser': 2.45.4
-  '@amplitude/plugin-session-replay-browser': 1.33.6
+  '@amplitude/analytics-browser': 2.45.5
+  '@amplitude/plugin-session-replay-browser': 1.33.7
   '@base-ui/react': 1.7.0
   '@chromatic-com/storybook': 5.2.1
-  '@cucumber/cucumber': 13.2.0
+  '@cucumber/cucumber': 13.2.1
   '@egoist/tailwindcss-icons': 1.9.2
   '@emoji-mart/data': 1.2.1
   '@eslint-community/eslint-plugin-eslint-comments': 4.7.2
-  '@eslint-react/eslint-plugin': 5.18.0
+  '@eslint-react/eslint-plugin': 5.18.2
   '@eslint/markdown': 8.0.3
   '@floating-ui/react': 0.27.20
   '@formatjs/intl-localematcher': 0.8.13
   '@heroicons/react': 2.2.0
   '@hey-api/openapi-ts': 0.98.2
-  '@hono/node-server': 2.0.12
+  '@hono/node-server': 2.1.0
   '@iconify-json/heroicons': 1.2.3
   '@iconify-json/ri': 1.2.10
   '@lexical/code': 0.47.0
@@ -77,45 +77,45 @@ catalog:
   '@mdx-js/loader': 3.1.1
   '@mdx-js/react': 3.1.1
   '@mdx-js/rollup': 3.1.1
-  '@mediabunny/mp3-encoder': 1.51.0
+  '@mediabunny/mp3-encoder': 1.52.3
   '@monaco-editor/react': 4.7.0
   '@napi-rs/keyring': 1.3.0
   '@next/mdx': 16.3.0
-  '@orpc/client': 1.14.10
-  '@orpc/contract': 1.14.10
-  '@orpc/openapi-client': 1.14.10
-  '@orpc/tanstack-query': 1.14.10
-  '@playwright/test': 1.62.0
+  '@orpc/client': 1.14.14
+  '@orpc/contract': 1.14.14
+  '@orpc/openapi-client': 1.14.14
+  '@orpc/tanstack-query': 1.14.14
+  '@playwright/test': 1.62.1
   '@remixicon/react': 4.9.0
   '@rgrove/parse-xml': 4.2.3
-  '@sentry/react': 10.68.0
-  '@storybook/addon-a11y': 10.5.4
-  '@storybook/addon-docs': 10.5.4
-  '@storybook/addon-links': 10.5.4
-  '@storybook/addon-onboarding': 10.5.4
-  '@storybook/addon-themes': 10.5.4
-  '@storybook/addon-vitest': 10.5.4
-  '@storybook/nextjs-vite': 10.5.4
-  '@storybook/react': 10.5.4
-  '@storybook/react-vite': 10.5.4
+  '@sentry/react': 10.69.0
+  '@storybook/addon-a11y': 10.5.6
+  '@storybook/addon-docs': 10.5.6
+  '@storybook/addon-links': 10.5.6
+  '@storybook/addon-onboarding': 10.5.6
+  '@storybook/addon-themes': 10.5.6
+  '@storybook/addon-vitest': 10.5.6
+  '@storybook/nextjs-vite': 10.5.6
+  '@storybook/react': 10.5.6
+  '@storybook/react-vite': 10.5.6
   '@streamdown/math': 1.0.2
-  '@svgdotjs/svg.js': 3.2.7
+  '@svgdotjs/svg.js': 3.2.8
   '@t3-oss/env-core': 0.13.11
   '@t3-oss/env-nextjs': 0.13.11
   '@tailwindcss/postcss': 4.3.3
   '@tailwindcss/typography': 0.5.20
   '@tailwindcss/vite': 4.3.3
   '@tanstack/eslint-plugin-query': 5.101.4
-  '@tanstack/form-core': 1.33.2
+  '@tanstack/form-core': 1.33.3
   '@tanstack/query-core': 5.101.4
-  '@tanstack/react-form': 1.33.2
+  '@tanstack/react-form': 1.33.3
   '@tanstack/react-hotkeys': 0.10.0
   '@tanstack/react-query': 5.101.4
-  '@tanstack/react-virtual': 3.14.8
+  '@tanstack/react-virtual': 3.14.9
   '@testing-library/dom': 10.4.1
   '@testing-library/jest-dom': 6.9.1
   '@testing-library/react': 16.3.2
-  '@testing-library/user-event': 14.6.1
+  '@testing-library/user-event': 14.6.3
   '@tsslint/cli': 3.1.4
   '@tsslint/compat-eslint': 3.1.4
   '@tsslint/config': 3.1.4
@@ -124,13 +124,13 @@ catalog:
   '@types/negotiator': 0.6.4
   '@types/node': 25.9.5
   '@types/qs': 6.15.1
-  '@types/react': 19.2.17
-  '@types/react-dom': 19.2.3
+  '@types/react': 19.2.18
+  '@types/react-dom': 19.2.4
   '@types/sortablejs': 1.15.9
-  '@typescript-eslint/parser': 8.65.0
+  '@typescript-eslint/parser': 8.66.0
   '@typescript/native': npm:typescript@7.0.2
-  '@vitejs/plugin-react': 6.0.4
-  '@vitejs/plugin-rsc': 0.5.30
+  '@vitejs/plugin-react': 6.0.5
+  '@vitejs/plugin-rsc': 0.5.33
   '@vitest/browser': 4.1.10
   '@vitest/browser-playwright': 4.1.10
   '@vitest/coverage-v8': 4.1.10
@@ -148,7 +148,7 @@ catalog:
   cron-parser: 5.6.2
   dayjs: 1.11.21
   decimal.js: 10.6.0
-  dompurify: 3.4.12
+  dompurify: 3.4.13
   echarts: 6.1.0
   echarts-for-react: 3.0.6
   elkjs: 0.11.1
@@ -164,59 +164,59 @@ catalog:
   eslint-plugin-command: 3.5.3
   eslint-plugin-erasable-syntax-only: 0.4.2
   eslint-plugin-hyoban: 0.14.1
-  eslint-plugin-jsdoc: 63.3.1
-  eslint-plugin-jsonc: 3.3.0
+  eslint-plugin-jsdoc: 63.3.3
+  eslint-plugin-jsonc: 3.4.1
   eslint-plugin-markdown-preferences: 0.41.1
   eslint-plugin-n: 18.2.2
   eslint-plugin-no-barrel-files: 1.3.1
-  eslint-plugin-perfectionist: 5.10.0
+  eslint-plugin-perfectionist: 5.10.1
   eslint-plugin-pnpm: 1.7.0
   eslint-plugin-regexp: 3.1.1
-  eslint-plugin-storybook: 10.5.4
+  eslint-plugin-storybook: 10.5.6
   eslint-plugin-toml: 1.5.0
   eslint-plugin-unicorn: 71.1.0
-  eslint-plugin-yml: 3.6.0
+  eslint-plugin-yml: 3.8.1
   eventsource-parser: 3.1.0
   fast-deep-equal: 3.1.3
   foxact: 0.3.8
   fuse.js: 7.5.0
   happy-dom: 20.11.1
   hast-util-to-jsx-runtime: 2.3.6
-  hono: 4.12.32
+  hono: 4.13.0
   html-entities: 2.6.0
   html-to-image: 1.11.13
   i18next: 26.3.6
-  i18next-resources-to-backend: 1.2.1
+  i18next-resources-to-backend: 1.2.3
   iconify-import-svg: 0.2.0
-  immer: 11.1.15
+  immer: 11.1.16
   jotai: 2.20.2
   jotai-effect: 2.4.1
   jotai-scope: 0.11.0
   jotai-tanstack-query: 0.11.0
   js-cookie: 3.0.8
-  js-yaml: 5.2.2
+  js-yaml: 5.2.3
   jsonschema: 1.5.0
   katex: 0.17.0
-  knip: 6.29.0
+  knip: 6.31.0
   ky: 2.0.2
   lexical: 0.47.0
   lockfile: 1.0.4
-  loro-crdt: 1.13.8
-  mediabunny: 1.51.0
-  mermaid: 11.16.0
+  loro-crdt: 1.13.9
+  mediabunny: 1.52.3
+  mermaid: 11.16.1
   mime: 4.1.0
   mitt: 3.0.1
-  motion: 12.42.2
+  motion: 12.43.0
   negotiator: 1.0.0
   next: 16.3.0
   next-themes: 0.4.6
-  nuqs: 2.9.2
+  nuqs: 2.9.5
   open: 11.0.0
   ora: 9.4.1
   picocolors: 1.1.1
   pinyin-pro: 3.28.2
-  playwright: 1.62.0
-  postcss: 8.5.23
+  playwright: 1.62.1
+  postcss: 8.5.25
   qrcode.react: 4.2.0
   qs: 6.15.3
   react: 19.2.8
@@ -234,17 +234,17 @@ catalog:
   scheduler: 0.27.0
   server-only: 0.0.1
   sharp: 0.35.3
-  shiki: 4.3.1
+  shiki: 4.4.2
   socket.io-client: 4.8.3
   sortablejs: 1.15.7
   std-semver: 1.0.8
-  storybook: 10.5.4
+  storybook: 10.5.6
   streamdown: 2.5.0
   string-ts: 2.3.1
   tailwind-merge: 3.6.0
   tailwindcss: 4.3.3
-  tldts: 7.4.9
-  tsx: 4.23.1
+  tldts: 7.4.10
+  tsx: 4.23.8
   typescript: npm:@typescript/typescript6@6.0.2
   uglify-js: 3.19.3
   undici: 7.29.0
@@ -252,9 +252,9 @@ catalog:
   use-context-selector: 2.0.0
   uuid: 14.0.1
   vinext: 1.0.0-beta.4
-  vite: npm:@voidzero-dev/vite-plus-core@0.2.6
+  vite: npm:@voidzero-dev/vite-plus-core@0.2.8
   vite-plugin-inspect: 12.0.2
-  vite-plus: 0.2.6
+  vite-plus: 0.2.8
   vitest: 4.1.10
   vitest-browser-react: 2.2.0
   vitest-canvas-mock: 1.1.4
